+++ src/Planestandar.java
@@ -1,5 +1,5 @@
 public class Planestandar implements PlanSuscripcion {
     public double calcularCosto(int meses){
-        return meses*9;
+        return (double) meses*9;
     }
 }
